Multiple issue commit.
references #10 Remove dependency on Cocoon gem.
references #11 Allow schemas to be searched on the sechemas index page.
references #12 Allow schemas to be sorted with clickable headers on the schemas index page.
references #13 Add navigation to schemas in the main navigation menu.
references #14 Allow object values to be searched on the object values index page.
references #15 Allow object values to be sorted with clickable headers on the object values index page.